requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
google-generativeai>=0.3.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
asyncio-mqtt>=0.16.1
lxml>=4.9.3
